+++ src/generate-gtfs-derived-assets.js
@@ -2,7 +2,7 @@ const { createReadStream, readFileSync } = require('node:fs');
 const { writeFile, mkdtemp } = require('node:fs/promises');
 const turfConvex = require('@turf/convex').default;
 const turfBuffer = require('@turf/buffer').default;
-const { flattenStopRoutes } = require('./flatten-stop-routes');
+const { flattenStopRoutes } = require('./stop-and-route-lookup-helper');
 const { getDistanceAlongLookup } = require('./distance-along-lookup-helper');
 const { resolve, join } = require("path");
 const { tmpdir } = require('node:os');
@@ -10,13 +10,14 @@ const { filterRouteIds, filterTripIds, getInterestingStopIds, getInterestingStop
 const { parse } = require('csv-parse');
 
 /*
- * This script computes a polygon to define the "transit service area". The
- * purpose for this is, if your instance supports streets routing over a wider
- * geographical area than you have local transit information for, to warn your
- * user if local transit options relevant to their journey might be missing.
- *
- * The approach is to compute a buffered hull around all the transit stops,
- * excluding some stops that are filtered out by route ID or agency ID.
+ * This script generates two assets from the GTFS zip file. 
+ * These assets are used in bikehopper-web-app to expose some data from the GTFS files.
+ * The assets are:
+ *  1. transit-service-area.json: 
+ *     A rough GeoJSON polygon describing the area served transit in the GTFS
+ *  2. route-line-lookup.json
+ *     Lookup tables that provide easy lookups for locations of transit stops, 
+ *     route LineString shapes, and extra information for clipping route LineStrings between two stops.
  */
 
 const requiredGTFSFiles = new Set(['routes.txt', 'trips.txt', 'stop_times.txt', 'stops.txt']);
@@ -24,7 +25,18 @@ const ENV_FILTERED_AGENCY_IDS = process.env.FILTERED_AGENCY_IDS || '';
 const ENV_MANUALLY_FILTERED_ROUTE_IDS = process.env.MANUALLY_FILTERED_ROUTE_IDS || '';
 
 (async () => {
-  // PART 1: Computing transit-service-area.json
+  /* 
+   * PART 1: Computing transit-service-area.json
+   *
+   * This part computes a polygon to define the "transit service area". The
+   * purpose for this is, if your instance supports streets routing over a wider
+   * geographical area than you have local transit information for, to warn your
+   * user if local transit options relevant to their journey might be missing.
+   *
+   * The approach is to compute a buffered hull around all the transit stops,
+   * excluding some stops that are filtered out by route ID or agency ID.
+   */
+
 
   // Initialize temprary folders to hold gtfs files
   const gtfsFilePath = resolve(process.env.GTFS_ZIP_PATH);
@@ -83,16 +95,31 @@ const ENV_MANUALLY_FILTERED_ROUTE_IDS = process.env.MANUALLY_FILTERED_ROUTE_IDS 
   
   console.log(`Finished writing transit-service-area.json to: ${outputPath}`)
 
-  // PART 2: Computing lookup tables for transit route geometry
+   /* 
+   * PART 2: Computing route-line-lookup.json
+   *
+   * This part computes three lookup tables:
+   * 1. stopIdPointLookup: 
+   *    Key is the stop-id of the stop, and the Value is a GeoJSON point of the stops location
+   * 2. routeIdLineStringLookup:
+   *    Key is the route-id of a route, and the value is a GeoJSON LineString of the entire route
+   * 3. distanceAlongLookup:
+   *    This is a two-level dictionary
+   *    Level1 :
+   *    Key is a stop-id, Value is the 2nd Level dictionary
+   *       Level 2:
+   *       Key is a trip-id, Value is how far along the LineString the stop is for that trip.
+   *
+   * These three lookup tables frovide enough information to generate a LineString for a
+   * trip thats clipped between the entry and exit stops. 
+   */
   const gtfsToGeoJSON = await import('gtfs-to-geojson');
-
   const agencies = [
     {
       agency_key: 'RG',
       path: gtfsFilePath,
     }
   ];
-  
   await gtfsToGeoJSON.default({
     agencies,
     outputType: 'agency',
@@ -100,15 +127,17 @@ const ENV_MANUALLY_FILTERED_ROUTE_IDS = process.env.MANUALLY_FILTERED_ROUTE_IDS 
     ignoreDuplicates: true,
   });
 
-  // This is a hardcoded temp geojson file thats an intermediary
+  // 'gtfs-to-geojson' library generates the geojson from the GTFS into this path, we read it in here
   const outputGeojsonPath = join(process.cwd(), '/geojson/RG/RG.geojson');
-  
   const geojson = JSON.parse(readFileSync(outputGeojsonPath, {encoding: 'utf-8'}));
-  const {stopIdPointLookup, routeIdLineStringLookup} = flattenStopRoutes(geojson);
+
+  // Lookup tables for stop-points and route linestrings
+  const {stopIdPointLookup, routeIdLineStringLookup} = getStopAndRouteLookups(geojson);
 
   stopTimesReadableStream = createReadStream(resolve(gtfsOutputPath, `stop_times.txt`), {encoding: 'utf8'});
   const parser = stopTimesReadableStream.pipe(parse());
 
+  // Lookup tables for clipping LineStrings
   const distanceAlongLookup = await getDistanceAlongLookup(parser);
   
   const routlineLookups = {
